requirements.txt
setuptools
django
djangorestframework
firebase-admin>=6.5.0
supabase